requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
click==8.1.7
Django==4.2.16
django-postgres-extra==2.0.8
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
psycopg==3.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
